+++ src/modules/scripting/dragonscript/src/classes/canvas/deClassCanvasCanvasView.cpp
@@ -164,7 +164,7 @@ dsFunction( init.clsCCView, "hashCode", DSFT_FUNCTION, DSTM_PUBLIC | DSTM_NATIVE
 void deClassCanvasCanvasView::nfHashCode::RunFunction( dsRunTime *rt, dsValue *myself ){
 	deCanvasCanvasView * const canvas = ( ( sCCViewNatDat* )p_GetNativeData( myself ) )->canvas;
 	// hash code = memory location
-	rt->PushInt( ( intptr_t )canvas );
+	rt->PushInt( ( int )( intptr_t )canvas );
 }
 
 // public func bool equals( Object obj )
