+++ src/modules/scripting/dragonscript/src/classes/canvas/deClassCanvasImage.cpp
@@ -162,7 +162,7 @@ dsFunction( init.clsCImage, "hashCode", DSFT_FUNCTION, DSTM_PUBLIC | DSTM_NATIVE
 void deClassCanvasImage::nfHashCode::RunFunction( dsRunTime *rt, dsValue *myself ){
 	deCanvasImage * const canvas = ( ( sCImageNatDat* )p_GetNativeData( myself ) )->canvas;
 	// hash code = memory location
-	rt->PushInt( ( intptr_t )canvas );
+	rt->PushInt( ( int )( intptr_t )canvas );
 }
 
 // public func bool equals( Object obj )
