+++ src/modules/scripting/dragonscript/src/classes/collider/deClassCollider.cpp
@@ -1868,7 +1868,7 @@ dsFunction( init.clsCol, "hashCode", DSFT_FUNCTION, DSTM_PUBLIC | DSTM_NATIVE, i
 void deClassCollider::nfHashCode::RunFunction( dsRunTime *rt, dsValue *myself ){
 	deCollider *collider = ( ( sColNatDat* )p_GetNativeData( myself ) )->collider;
 	// hash code = memory location
-	rt->PushInt( ( intptr_t )collider );
+	rt->PushInt( ( int )( intptr_t )collider );
 }
 
 // public func bool equals( Object obj )
