+++ src/modules/scripting/dragonscript/src/classes/sound/deClassMicrophone.cpp
@@ -342,7 +342,7 @@ dsFunction( init.clsMic, "hashCode", DSFT_FUNCTION, DSTM_PUBLIC | DSTM_NATIVE, i
 void deClassMicrophone::nfHashCode::RunFunction( dsRunTime *rt, dsValue *myself ){
 	deMicrophone *microphone = ( ( sMicNatDat* )p_GetNativeData( myself ) )->microphone;
 	
-	rt->PushInt( ( intptr_t )microphone );
+	rt->PushInt( ( int )( intptr_t )microphone );
 }
 
 // public func bool equals( Object object )
